I managed to get the logic right this time
We generate the first guess
We store the correct indexes for the right colors so this can help the computer chose from a smaller pool each time
We then save the colors on the array and they remain unchanged if they are got right
Each time we generate a new guess,we omit the correct indexes that we got in the previous guess
Whenever we find a color that is not on the right position but is part of the array,we store it
These colors are saved for the next guesses

I used my own tactic for the computer guess and it goes like this:
    if a color from the guess is not even a part of the player's choice,
    we remove it from the pool of colors
    this way the computer doesn't waste guesses on colors that are not even part of the players's pick

Thats about it, hope you enjoyed this! If there are any improvements i should make,message me!